requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.3.0

pandas==2.2.2
xlrd==1.2.0
